Separated test for calc_psf_geometry
